Move Error C++ definitions to c++ file
