+++ v8go.cc
@@ -8,8 +8,6 @@
 
 #include <cstdlib>
 #include <cstring>
-#include <iostream>
-#include <sstream>
 #include <unordered_map>
 #include <vector>
 #include "utils.h"
@@ -28,46 +26,6 @@ struct m_template {
   Isolate* iso;
   Persistent<Template> ptr;
 };
-
-RtnError ExceptionError(TryCatch& try_catch, Isolate* iso, Local<Context> ctx) {
-  HandleScope handle_scope(iso);
-
-  RtnError rtn = {nullptr, nullptr, nullptr};
-
-  if (try_catch.HasTerminated()) {
-    rtn.msg =
-        CopyString("ExecutionTerminated: script execution has been terminated");
-    return rtn;
-  }
-
-  String::Utf8Value exception(iso, try_catch.Exception());
-  rtn.msg = CopyString(exception);
-
-  Local<Message> msg = try_catch.Message();
-  if (!msg.IsEmpty()) {
-    String::Utf8Value origin(iso, msg->GetScriptOrigin().ResourceName());
-    std::ostringstream sb;
-    sb << *origin;
-    Maybe<int> line = try_catch.Message()->GetLineNumber(ctx);
-    if (line.IsJust()) {
-      sb << ":" << line.ToChecked();
-    }
-    Maybe<int> start = try_catch.Message()->GetStartColumn(ctx);
-    if (start.IsJust()) {
-      sb << ":"
-         << start.ToChecked() + 1;  // + 1 to match output from stack trace
-    }
-    rtn.location = CopyString(sb.str());
-  }
-
-  Local<Value> mstack;
-  if (try_catch.StackTrace(ctx).ToLocal(&mstack)) {
-    String::Utf8Value stack(iso, mstack);
-    rtn.stack = CopyString(stack);
-  }
-
-  return rtn;
-}
 
 m_unboundScript* tracked_unbound_script(m_ctx* ctx, m_unboundScript* us) {
   ctx->unboundScripts.push_back(us);
